% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\alias{print.boot.glmnet}
\title{Bootstrap Validation for Survival Model}
\usage{
bootstrap(
  x,
  y,
  fun = rcv.glmnet,
  nboot = 200L,
  m = 50,
  times = 90,
  ...,
  s = "lambda.1se",
  verbose = interactive()
)

\method{print}{boot.glmnet}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{\code{matrix}, data/feature matrix.}

\item{y}{\code{Surv}, survival time and status as \code{Surv} object.}

\item{fun}{model function, e.g. \code{\link[=rcv.glmnet]{rcv.glmnet()}}.}

\item{nboot}{\code{integer} number of bootstrap samples}

\item{m}{\code{integer}, individuals/observations per interval}

\item{times}{\code{numeric} predict survival at \code{times}.}

\item{s}{\code{character}/\code{numeric}, value(s) of the penality parameter \code{lambda}.
See \code{\link[glmnet:predict.cv.glmnet]{glmnet::predict.cv.glmnet()}} for details.}

\item{verbose}{\code{logical}, if \code{TRUE} a progressbar is shown.}

\item{digits}{\code{integer(1)}, number of digits shown in table.}

\item{\ldots}{further params passed to \code{fun}.}
}
\value{
A \code{list}, with the fitted model \code{fit} and the over-optimistic error.
}
\description{
Boostrap validation for survival data as described in Harrell et al. 1996.
}
\examples{
# nboot should usually be higher but to keep the runtime of the example low
# we choose 2 here
data(eldd)
x <- na.omit(eldd)
y <- Surv(x$DaysAtRisk, x$Deceased)
x <- as.matrix(x[,c("Age", "ALB_S", "BILI_S", "CRE_S", "INR_C")])
boot <- bootstrap(
    x, y, rcv.glmnet, family = "cox",
    nboot = 2, nrepcv = 2, nfolds = 3
)
boot
}
\references{
Harrell Jr, Frank E., Kerry L. Lee, and Daniel B. Mark.
"Multivariable prognostic models: issues in developing models, evaluating
assumptions and adequacy, and measuring and reducing errors."
Statistics in medicine 15.4 (1996): 361-387.
\doi{10.1002/(SICI)1097-0258(19960229)15:4<361::AID-SIM168>3.0.CO;2-4}
}
