% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-timeroc.R
\name{confidencebands}
\alias{confidencebands}
\title{Confidence bands}
\usage{
confidencebands(
  T,
  delta,
  marker,
  timepoint,
  cause = 1,
  ...,
  nboot = 200,
  conf.level = 0.95,
  verbose = interactive()
)
}
\arguments{
\item{T}{\code{numeric}, time, see \code{\link[timeROC:timeROC]{timeROC::timeROC()}}.}

\item{delta}{\code{numeric}, status, see \code{\link[timeROC:timeROC]{timeROC::timeROC()}}.}

\item{marker}{\code{numeric}, marker, see \code{\link[timeROC:timeROC]{timeROC::timeROC()}}.}

\item{timepoint}{\code{numeric(1)}, timepoint to evaluate.}

\item{cause}{\code{numeric}, cause, see \code{\link[timeROC:timeROC]{timeROC::timeROC()}}.}

\item{nboot}{\code{numeric(1)}, number of boostrap samples.}

\item{conf.level}{\code{numeric(1)}, confidence level.}

\item{verbose}{\code{logical}, if \code{TRUE} a progressbar is shown.}

\item{\ldots}{further arguments passed to \code{timeROC}.}
}
\value{
a \code{list} with 2 elements (TP, FP), each one a \code{matrix},
with 3 rows (lower/median/upper) band, and  \code{length(threshold)} columns.
}
\description{
Calculate confidence bands for a single timepoint timeROC ROC
based on threshold averaging as described in Fawcett 2004.
}
\references{
Fawcett, T. (2004).
ROC graphs: Notes and practical considerations for researchers.
Machine learning, 31(1), 1-38.
}
