% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eldr}
\alias{eldr}
\title{End-stage Liver Disease Reference}
\format{
A \code{data.frame} with 36 rows and 8 columns. Each row describes a
laboratory measurement/method.
\describe{
\item{Code}{\code{character}, name.}
\item{Unit}{\code{character}, unit.}
\item{LongDescription}{\code{character}, full name.}
\item{ShortDescription}{\code{character}, abbreviation.}
\item{LowerLimit}{\code{numeric}, lower limit.}
\item{UpperLimit}{\code{numeric}, upper limit.}
\item{AgeDays}{\code{integer}, the method is only valid for patients older than
\code{AgeDays} days.}
\item{Sex}{\code{factor}, the method is only valid for "male" or "female"
patients. Or the sex doesn't matter ("both").}
}
}
\source{
Institute of Laboratory Medicine,
Clinical Chemistry and Molecular Diagnostics,
University Hospital Leipzig,
Paul-List-Str. 13-15, D-04103 Leipzig, Germany.
}
\usage{
eldr
}
\description{
A dataset of reference limits of the University Hospital Leipzig for
laboratory measurements used in \code{\link{eldd}}.
}
\keyword{datasets}
