% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-text.R
\name{enum}
\alias{enum}
\title{Format enumeration}
\usage{
enum(..., conjunction = "and", oxford = TRUE)
}
\arguments{
\item{...}{one or more objects, converted to characters}

\item{conjunction}{\code{character}, final conjunction}

\item{oxford}{\code{logical}, oxford comma?}
}
\description{
Helper function to format an enumeration.
}
\examples{
enum(letters[1:2])
enum(letters[1:3])
}
