% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meld.R
\name{meld}
\alias{meld}
\alias{meld_na}
\alias{meld_plus7}
\title{Calculate MELD scores}
\usage{
meld(
  creatinine,
  bilirubin,
  inr,
  dialysis = FALSE,
  cause = c("other", "unos", "ethyltoxic", "cholestatic"),
  round = FALSE
)

meld_na(
  creatinine,
  bilirubin,
  inr,
  sodium,
  dialysis = FALSE,
  type = c("kim2008", "unos"),
  cause = "other",
  round = FALSE
)

meld_plus7(
  creatinine,
  bilirubin,
  inr,
  sodium,
  albumin,
  wbc,
  age,
  round = FALSE
)
}
\arguments{
\item{creatinine}{\code{numeric}, [mg/dl]}

\item{bilirubin}{\code{numeric}, [mg/dl]}

\item{inr}{\code{numeric}}

\item{dialysis}{\code{logical}, had dialysis twice, or 24 hours of CVVHD, within
a week prior to the serum creatinine test?}

\item{cause}{\code{character}, cause of cirrhosis. Use \code{"unos"} for the United
Network for Organ Sharing definition of the score.}

\item{round}{\code{logical}, round to nearest integer?}

\item{sodium}{\code{numeric}, [mmol/l]}

\item{type}{\code{character(1)}, type of implementation}

\item{albumin}{\code{numeric}, [g/dl]}

\item{wbc}{\code{numeric}, [Gpt/l]}

\item{age}{\code{numeric}, [years]}
}
\value{
\code{numeric}
}
\description{
Calculate MELD, MELD-Na and MELD-Plus scores.
}
\details{
Laboratory values (creatinine, bilirubin, INR) below 1.0 are set to 1.0 and
creatinine above 4.0 is set to 4.0.

In contrast to \code{meld} and \code{meldNa}, \code{meldPlus7} returns a probability.
}
\examples{
meld(creatinine = 1.9, bilirubin = 4.2, inr = 1.2, cause = "other")

meld_na(creatinine = 1.9, bilirubin = 4.2, inr = 1.2, sodium = 135)
meld_na(creatinine = 1.9, bilirubin = 4.2, inr = 1.2, sodium = 135,
        type = "unos")
meld_plus7(2.5, 4.1, 1.2, 137, 24, 6.7, 56)
}
\references{
Michael Malinchoc et al. 2000.
"A model to predict poor survival in patients undergoing transjugular
intrahepatic portosystemic shunts"
Hepatology 31 (4): 864-871.
\doi{10.1053/he.2000.5852}

United Network for Organ Sharing (UNOS). Policy notice 11/2015:
OPTN executive committee actions. Available at:
https://optn.transplant.hrsa.gov/media/1575/policynotice_20151101.pdf
Accessed: September 11, 2019.

Kim et al. 2008.
"Hyponatremia and Mortality among Patients on the Liver-Transplant Waiting
List"
N Engl J Med 2008; 359: 1018-1026
\doi{10.1056/NEJMoa0801209}

Kartoun et al. 2017.
"The MELD-Plus: A generalizable prediction risk score in cirrhosis"
PloS one, 12 (10): e0186301
\doi{10.1371/journal.pone.0186301}
}
