% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-survival.R
\name{observed_survival}
\alias{observed_survival}
\alias{observed_mortality}
\alias{observed_events}
\title{Calculate observed survival informations}
\usage{
observed_survival(x, f, times)

observed_mortality(x, f, times)

observed_events(x, f, times)
}
\arguments{
\item{x}{\code{Surv}, object}

\item{f}{\code{factor}, strata/splitting factor}

\item{times}{\code{numeric}, information at times}
}
\value{
\code{double}, with an element for each level in \code{f},
if times is a \code{numeric} of length 1;  or a \code{matrix} with
as much rows as elements in \code{times} and a column for each level in \code{f}.
}
\description{
This functions calculate some survival information (e.g. survival rate,
mortality rate).
}
\details{
In contrast to \code{\link[survival:summary.survfit]{survival::summary.survfit()}} these functions do not remove
empty strata (e.g. important for bootstrapping).
}
\examples{
sv <- Surv(1:4, c(0, 1, 1, 0))
f <- factor(c("a", "c", "a", "c"), levels = c("a", "b", "c"))
observed_survival(sv, times = 3)
observed_survival(sv, f = f, times = 3)
observed_survival(sv, f = f, times = 1:4)

observed_mortality(sv, f = f, times = 3)

observed_events(sv, f = f, times = 3)
}
