% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-survival.R
\name{observed_vs_expected_mortality}
\alias{observed_vs_expected_mortality}
\title{Calculate observed vs expected mortality}
\usage{
observed_vs_expected_mortality(x, f, time, expected)
}
\arguments{
\item{x}{\code{Surv}, object}

\item{f}{\code{factor}, strata/splitting factor}

\item{time}{\code{numeric(1)}, observed/expected survival at \code{time} timepoint.}

\item{expected}{\code{numeric}, expected mortality rate, has to be of the same
length as levels in \code{f}.}
}
\value{
\code{data.frame}, with observed/expected mortality rates and events.
}
\description{
This function creates a table to compare the observed and the expected
mortality.
}
\examples{
sv <- Surv(1:8, c(0, 1, 1, 0, 1, 1, 1, 0))
f <- factor(rep(c("a", "c"), 4), levels = c("a", "b", "c"))
observed_vs_expected_mortality(sv, f, time = 3, expected = c(0.5, 0.3, 0.2))
}
