% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcv.glmnet.R
\name{plot.arcv.glmnet}
\alias{plot.arcv.glmnet}
\title{Plot the elastic net cross-validation curves}
\usage{
\method{plot}{arcv.glmnet}(
  x,
  col = viridisLite::cividis(length(x$alpha)),
  what = c("all", "lambda.min", "lambda.1se"),
  main = NULL,
  pch = 20L,
  ...
)
}
\arguments{
\item{x}{\code{arcv.glmnet} object.}

\item{col}{\code{character/numeric}, colours.}

\item{what}{\code{character(1)}, what to plot: \code{"all"} plot all cross-validated
loss errors vs lambda of all alpha values, \code{"lambda.min"}/\code{"lambda.1se"}
plots the "best" lambda for each alpha.}

\item{main}{\code{character(1)}, title.}

\item{pch}{\code{character/numeric}, point character/symbol.}

\item{\dots}{further arguments passed to \code{plot}.}
}
\description{
This functions plots the aggregated cross-validation curves produced by
\code{\link[=arcv.glmnet]{arcv.glmnet()}}.
}
\seealso{
\code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}}
}
\author{
Sebastian Gibb
}
