% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{plot.boot.glmnet}
\alias{plot.boot.glmnet}
\title{Plot the bootstrap results.}
\usage{
\method{plot}{boot.glmnet}(
  x,
  col = head(viridisLite::viridis(3L), 2L),
  what = c("calibration", "selected"),
  pch = 19L,
  ...
)
}
\arguments{
\item{x}{\code{boot.glmnet} object.}

\item{col}{\code{character/numeric}, colours.}

\item{what}{\code{character(1)}, what to plot: \code{"selected"} select variables.}

\item{pch}{\code{character/numeric}, point character/symbol.}

\item{\dots}{further arguments passed to \code{plot}.}
}
\description{
This functions plots the bootstrap results.
}
\seealso{
\code{\link[=rcv.glmnet]{rcv.glmnet()}}, \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}}
}
\author{
Sebastian Gibb
}
