% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dots}
\alias{plot_dots}
\title{Plot dotchart}
\usage{
plot_dots(
  x,
  xlim = c(0, max(x)),
  main = "Dotchart",
  xlab = "Frequency",
  col = palette.colors(2L)[2L],
  pch = 19L
)
}
\arguments{
\item{x}{\code{numeric}, values to plot.}

\item{xlim}{\code{numeric(2)}, limits of the x-axis.}

\item{main}{\code{character(1)}, plot title.}

\item{xlab}{\code{character(1)}, x-axis label.}

\item{col}{\code{integer}/\code{character}, color of the dots.}

\item{pch}{\code{integer}/\code{character}, point character/symbol of the dots.}
}
\value{
nothing, used for its side-effects (plotting).
}
\description{
This is a simplified and customized version of \code{\link[graphics:dotchart]{graphics::dotchart()}}.
}
\examples{
x <- c(Foo = 3, Bar = 5)
plot_dots(x, xlim = c(0, 8))
}
\seealso{
\code{\link[graphics:dotchart]{graphics::dotchart()}}
}
