% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_surv}
\alias{plot_surv}
\title{Plot method for 'survfit' objects}
\usage{
plot_surv(
  x,
  main = character(),
  xlab = "Time",
  ylab = "Overall survival probability",
  mark.time = TRUE,
  conf.int = FALSE,
  col = palette.colors(max(1L, length(x$strata))),
  times,
  ...
)
}
\arguments{
\item{x}{\code{survfit} object.}

\item{main}{\code{character(1)}, plot title.}

\item{xlab}{\code{character(1)}, x-axis label.}

\item{ylab}{\code{character(1)}, y-axis label.}

\item{mark.time}{\code{logical(1)}, if \code{TRUE} censoring times are marked, see
\code{\link[survival:plot.survfit]{survival::plot.survfit()}} for details.}

\item{conf.int}{\code{logical(1)}, if \code{TRUE} confidence interval is plotted, see
\code{\link[survival:plot.survfit]{survival::plot.survfit()}} for details.}

\item{col}{\code{integer}/\code{character}, specifying colors for each curve.}

\item{times}{\code{integer}, vector of times to print on the x-axis.}

\item{\dots}{further arguments passed to \code{\link[survival:plot.survfit]{survival::plot.survfit()}}.}
}
\value{
a list with \code{x} and \code{y} containing the coordinates of the last point
of each curves.
}
\description{
This is just a wrapper method around \code{\link[survival:plot.survfit]{survival::plot.survfit()}} with
custom defaults.
}
\examples{
library("survival")
srvfit <- survfit(Surv(time, status) ~ x, data = aml)
plot_surv(srvfit)
}
\seealso{
\code{\link[survival:plot.survfit]{survival::plot.survfit()}}
}
