% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_table}
\alias{plot_table}
\title{Plot a table}
\usage{
plot_table(
  x,
  main = character(),
  xlab = character(),
  ylab = character(),
  at = seq_len(nrow(x)) - 1L,
  xlim = range(at),
  ylim = c(0L, ncol(x)),
  ylabels = TRUE,
  col = rep_len(1L, ncol(x)),
  xaxis = TRUE,
  cex.xaxis = 3/4,
  cex.yaxis = 1.25,
  cex.text = 1.5,
  ...
)
}
\arguments{
\item{x}{\code{matrix}, it is transposed on the graphic device. The column names
correspond to the y labels and the row names to the x labels.}

\item{main}{\code{character(1)}, plot title.}

\item{xlab}{\code{character(1)}, x-axis label.}

\item{ylab}{\code{character(1)}, y-axis label.}

\item{at}{\verb{numeric, where to plot the rows of }x`.}

\item{xlim}{\code{numeric(2)}, limits of the x-axis.}

\item{ylim}{\code{numeric(2)}, limits of the y-axis.}

\item{ylabels}{\code{logical(1)}, should the column names used to labels the
y-axis (default: \code{TRUE})?}

\item{col}{\code{integer}/\code{character}, specifying the color for each y/column
label.}

\item{xaxis}{\code{logical(1)}, should the x-axis be plotted (default: \code{TRUE})?}

\item{cex.xaxis}{\code{numeric(1)}, character expansion factor for the x-axis
labels, see \code{\link[=par]{par()}} for details.}

\item{cex.yaxis}{\code{numeric(1)}, character expansion factor for the y-axis
labels, see \code{\link[=par]{par()}} for details.}

\item{cex.text}{\code{numeric(1)}, character expansion factor for the cell content
labels, see \code{\link[=par]{par()}} for details.}

\item{\dots}{further arguments passed to \code{\link[=plot.default]{plot.default()}}.}
}
\value{
nothing, used for its side-effects (plotting).
}
\description{
Plot a table on the current graphic device. Useful for risk tables.
}
\examples{
m <- matrix(
    1:8, nrow = 4,
    dimnames = list(c(0, 30, 90, 365), LETTERS[1:2])
)
plot_table(m, main = "Cumulative number of events")
}
