% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcv.glmnet.R
\name{predict.rcv.glmnet}
\alias{predict.rcv.glmnet}
\title{Predictions for a \code{rcv.glmnet} object}
\usage{
\method{predict}{rcv.glmnet}(
  object,
  newx,
  s = c("lambda.1se", "lambda.min"),
  type = c("link", "response", "coefficients", "nonzero", "class", "survival"),
  times,
  ...
)
}
\arguments{
\item{object}{\code{rcv.glmnet} object.}

\item{newx}{\code{matrix}, of new values for \code{x} at which predictions are to be
made.}

\item{s}{\code{character}/\code{numeric}, value(s) of the penality parameter \code{lambda}.
See \code{\link[glmnet:predict.cv.glmnet]{glmnet::predict.cv.glmnet()}} for details.}

\item{type}{\code{character}, type of prediction. For \code{"survival"}
the predicted survival is returned for each timepoint in \code{times}. All other
types are passed to \code{predict.cv.glmnet}
See \code{\link[glmnet:predict.cv.glmnet]{glmnet::predict.cv.glmnet()}} for details.}

\item{times}{\code{numeric}, vector of times, the returned matrix will contain one
row for each time. See \code{\link[survival:summary.survfit]{survival::summary.survfit()}} for details.}

\item{\dots}{further arguments passed to \code{predict.cv.glmnet}.}
}
\value{
The object returned depends on the \dots arguments.
See \code{\link[glmnet:predict.cv.glmnet]{glmnet::predict.cv.glmnet()}} for details. For \code{type = "survival"} the
returned value is a \code{matrix} with one row per \code{times} element and one column
for each row in \code{newx}.
}
\description{
Compute fitted values for a model fitted by \code{rcv.glmnet}.
}
\examples{
# Example adapted from ?"glmnet::cv.glmnet"
set.seed(10101)
n <- 500
p <- 30
nzc <- trunc(p / 10)
x <- matrix(rnorm(n * p), n, p)
beta <- rnorm(nzc)
fx <- x[, seq(nzc)] \%*\% beta / 3
hx <- exp(fx)
ty <- rexp(n, hx)
tcens <- rbinom(n = n, prob = 0.3, size = 1)  # censoring indicator
# y <- Surv(ty, 1-tcens) with library("survival")
y <- cbind(time = ty, status = 1 - tcens)
# nrepcv should usually be higher but to keep the runtime of the example low
# we choose 2 here
rcvob <- rcv.glmnet(x, y, family = "cox", nrepcv = 2, nfolds = 3)
predict(
    rcvob,
    newx = x[1:5,], x = x, y = survival::Surv(y[, "time"], y[, "status"]),
    type = "survival", times = c(0, 7), s = "lambda.1se"
)
}
\seealso{
\code{\link[=rcv.glmnet]{rcv.glmnet()}}, \code{\link[glmnet:predict.cv.glmnet]{glmnet::predict.cv.glmnet()}},
\code{\link[survival:summary.survfit]{survival::summary.survfit()}}
}
\author{
Sebastian Gibb
}
