% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rjlegend}
\alias{rjlegend}
\title{Legend with right justified text}
\usage{
rjlegend(x = "bottomright", y = NULL, legend, col, lwd = 1, bty = "n", ...)
}
\arguments{
\item{x}{position, see \code{\link[=legend]{legend()}}.}

\item{y}{position, see \code{\link[=legend]{legend()}}.}

\item{legend}{legend text, see \code{\link[=legend]{legend()}}.}

\item{col}{colours, see \code{\link[=legend]{legend()}}.}

\item{lwd}{line width, see \code{\link[=legend]{legend()}}.}

\item{bty}{border type, see \code{\link[=legend]{legend()}}.}

\item{...}{params passed to \code{\link[=legend]{legend()}}.}
}
\value{
see \code{\link[=legend]{legend()}}
}
\description{
Plots a legend similar to \code{\link[=legend]{legend()}} but with right justified text.
}
