% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si.R
\name{as_si}
\alias{as_si}
\alias{as_metric}
\title{Convert laboratory metric to SI units}
\usage{
as_si(x, type = c("bilirubin", "creatinine"))

as_metric(x, type = c("bilirubin", "creatinine"))
}
\arguments{
\item{x}{\code{numeric}, values to convert}

\item{type}{\code{character}, type of laboratory value}
}
\value{
\code{numeric}
}
\description{
Converts some laboratory metric to SI units.
}
\examples{
as_si(4.2, "creatinine")
as_metric(175, "creatinine")
}
