% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxnet.R
\name{basehaz}
\alias{basehaz}
\alias{basehaz.default}
\alias{basehaz.coxph}
\alias{basehaz.cv.glmnet}
\alias{basehaz.coxnet}
\title{Baseline hazard}
\usage{
basehaz(fit, ...)

\method{basehaz}{coxph}(fit, centered = TRUE, ...)

\method{basehaz}{cv.glmnet}(
  fit,
  x,
  y,
  s = c("lambda.1se", "lambda.min"),
  times = NULL,
  centered = TRUE,
  ...
)

\method{basehaz}{coxnet}(fit, x, y, s = NULL, times = NULL, centered = TRUE, ...)
}
\arguments{
\item{fit}{fitted model.}

\item{\dots}{further arguments passed to \code{basehaz}.}

\item{centered}{\code{logical(1)}, see \code{\link[survival:basehaz]{survival::basehaz()}}.}

\item{x}{\code{matrix}, model matrix used to fit \code{fit}.}

\item{y}{\code{Surv}, Survival object used to fit \code{fit}.}

\item{s}{\code{numeric(1)}, lambda penality parameter.}

\item{times}{\code{numeric}, estimate baseline hazard for times. If \code{NULL} all
times are returned.}
}
\value{
\code{data.frame} with variable names \code{hazard}, \code{time} and optionally
\code{strata}.
}
\description{
Make \verb{[survival::basehaz()]} available for \code{coxnet} objects.
}
\seealso{
\code{\link[survival:basehaz]{survival::basehaz()}}
}
\author{
Sebastian Gibb
}
