% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meld.R
\name{meld}
\alias{meld}
\alias{meld_na}
\alias{meld3}
\alias{pmeld}
\alias{pmeld3}
\alias{pmeld_plus7}
\title{Calculate MELD scores}
\usage{
meld(
  creatinine,
  bilirubin,
  inr,
  dialysis = FALSE,
  cause = "other",
  round = FALSE
)

meld_na(
  creatinine,
  bilirubin,
  inr,
  sodium,
  dialysis = FALSE,
  type = c("kim2008", "unos"),
  cause = "other",
  round = FALSE
)

meld3(
  creatinine,
  bilirubin,
  inr,
  sodium,
  albumin,
  female = TRUE,
  round = FALSE
)

pmeld(
  meld = NULL,
  creatinine,
  bilirubin,
  inr,
  dialysis = FALSE,
  cause = "other"
)

pmeld3(meld = NULL, creatinine, bilirubin, inr, sodium, albumin, female)

pmeld_plus7(
  creatinine,
  bilirubin,
  inr,
  sodium,
  albumin,
  wbc,
  age,
  round = FALSE
)
}
\arguments{
\item{creatinine}{\code{numeric}, [mg/dl]}

\item{bilirubin}{\code{numeric}, [mg/dl]}

\item{inr}{\code{numeric}}

\item{dialysis}{\code{logical}, had dialysis twice, or 24 hours of CVVHD, within
a week prior to the serum creatinine test?}

\item{cause}{\code{character}, cause of cirrhosis. Has to be of the same length as
\code{creatinine}, \code{bilirubin}, \code{inr} and \code{dialysis}. Just \code{"ethyltoxic"},
\code{"cholestatic"}, \code{"unos"} and \code{"other"} are supported.
Use \code{"unos"} for the United Network for Organ Sharing definition of the score.}

\item{round}{\code{logical}, round to nearest integer?}

\item{sodium}{\code{numeric}, [mmol/l]}

\item{type}{\code{character(1)}, type of implementation}

\item{albumin}{\code{numeric}, [g/dl]}

\item{female}{\code{logical}}

\item{meld}{\code{numeric}, MELD/MELD-Na score as calculated by \code{meld} or
\code{meld-na}.}

\item{wbc}{\code{numeric}, [Gpt/l]}

\item{age}{\code{numeric}, [years]}
}
\value{
\code{numeric}, MELD or MELD-Na score for \code{meld} and \code{meld_na}
respectively. \code{pmeld} and \code{pmeld_plus7} return the survival probability based
on the MELD and MELD-Plus7 score.
}
\description{
Calculate MELD, MELD-Na scores and MELD or MELD-Plus score based survival
probabilities.
}
\details{
Laboratory values (creatinine, bilirubin, INR) below 1.0 are set to 1.0 and
creatinine above 4.0 is set to 4.0 as described in Wiesner et al. 2003.

The original MELD-Plus risk score reports a mortality probability. To provide
a common interface with \code{pmeld} \code{pmeld_plus7} returns a survival probabilitiy
instead.
}
\examples{
meld(creatinine = 1.9, bilirubin = 4.2, inr = 1.2, cause = "other")

meld_na(creatinine = 1.9, bilirubin = 4.2, inr = 1.2, sodium = 135)
meld_na(creatinine = 1.9, bilirubin = 4.2, inr = 1.2, sodium = 135,
        type = "unos")

meld3(creatinine = 1.9, bilirubin = 4.2, inr = 1.2, sodium = 135, albumin = 3.5)

pmeld(creatinine = 1.9, bilirubin = 4.2, inr = 1.2, cause = "other")
pmeld(meld = 20)

pmeld3(meld = 20)

pmeld_plus7(2.5, 4.1, 1.2, 137, 24, 6.7, 56)
}
\references{
Michael Malinchoc et al. 2000.
"A model to predict poor survival in patients undergoing transjugular
intrahepatic portosystemic shunts"
Hepatology 31 (4): 864-871.
\doi{10.1053/he.2000.5852}

United Network for Organ Sharing (UNOS). Policy notice 11/2015:
OPTN executive committee actions. Available at:
https://optn.transplant.hrsa.gov/media/1575/policynotice_20151101.pdf
Accessed: September 11, 2019.

Kim et al. 2008.
"Hyponatremia and Mortality among Patients on the Liver-Transplant Waiting
List"
N Engl J Med 2008; 359: 1018-1026
\doi{10.1056/NEJMoa0801209}

Kim et al. 2021.
"MELD 3.0: The Model for End-Stage Liver Disease Updated for the Modern Era"
Gastroenterology 2021; 161: 1887-1895.e4
\doi{10.1053/j.gastro.2021.08.050}

Wiesner et al. 2003.
"Model for end-stage liver disease (MELD) and allocation of donor livers"
Gastroentrology, 124 (1): 91-96
\doi{10.1053/gast.2003.50016}

Kartoun et al. 2017.
"The MELD-Plus: A generalizable prediction risk score in cirrhosis"
PloS one, 12 (10): e0186301
\doi{10.1371/journal.pone.0186301}
}
