% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcv.glmnet.R
\name{plot.rcv.glmnet}
\alias{plot.rcv.glmnet}
\title{Plot the cross-validation curve}
\usage{
\method{plot}{rcv.glmnet}(x, ...)
}
\arguments{
\item{x}{\code{rcv.glmnet} object.}

\item{\dots}{further arguments passed to \code{plot.cv.glmnet}.}
}
\description{
This functions plots the aggregated cross-validation curve produced by
\code{\link[=rcv.glmnet]{rcv.glmnet()}}.
}
\seealso{
\code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}}
}
\author{
Sebastian Gibb
}
