% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_surv_roc}
\alias{plot_surv_roc}
\title{Plot survival ROC curves}
\usage{
plot_surv_roc(
  x,
  timepoint,
  col = setNames(palette.colors(length(x)), names(x)),
  lty = setNames(rep.int(1, length(x)), names(x)),
  main = paste0("ROC at day ", timepoint),
  xlab = "1 - Specificity",
  ylab = "Sensitivity"
)
}
\arguments{
\item{x}{\code{list} of \code{timeROC::ipcwsurvivalROC} objects.}

\item{timepoint}{\code{numeric(1)}, timepoints for ROC prediction}

\item{col}{\code{character}, colours}

\item{lty}{\code{numeric}, line type}

\item{main}{\code{character(1)}, title}

\item{xlab}{\code{character(1)}, label x-axis}

\item{ylab}{\code{character(1)}, label y-axis}
}
\value{
\code{double}, AUC
}
\description{
Generate ROC plots for a single timepoint for  \code{\link[timeROC:timeROC]{timeROC::timeROC()}} objects.
}
